requirements.txt
gradio==4.26.0
google-cloud-vision==3.10.2
pdf2image
pytesseract
Pillow

